requirements.txt
numpy>=1.24
pyadi-iio>=0.0.20
matplotlib>=3.8
scipy>=1.16
cryptography>=42.0
